Move internal theme classes to org.eclipse.tm4e.core.internal.theme
