+++ org.eclipse.tm4e.core/META-INF/MANIFEST.MF
@@ -20,6 +20,7 @@ Export-Package: org.eclipse.tm4e.core,
  org.eclipse.tm4e.core.grammar,
  org.eclipse.tm4e.core.internal.grammar;x-friends:="org.eclipse.tm4e.core.tests",
  org.eclipse.tm4e.core.internal.matcher;x-friends:="org.eclipse.tm4e.core.tests",
+ org.eclipse.tm4e.core.internal.theme;x-friends:="org.eclipse.tm4e.core.tests",
  org.eclipse.tm4e.core.internal.theme.reader;x-friends:="org.eclipse.tm4e.core.tests",
  org.eclipse.tm4e.core.internal.utils;x-friends:="org.eclipse.tm4e.core.tests,org.eclipse.tm4e.ui",
  org.eclipse.tm4e.core.model,
