+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/internal/grammar/StackElementMetadata.java
@@ -12,7 +12,7 @@
 package org.eclipse.tm4e.core.internal.grammar;
 
 import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.tm4e.core.theme.FontStyle;
+import org.eclipse.tm4e.core.internal.theme.FontStyle;
 
 /**
  * Metadata for {@link StackElement}.
