+++ org.eclipse.tm4e.core/src/main/java/org/eclipse/tm4e/core/internal/theme/ThemeRaw.java
@@ -16,9 +16,6 @@ import java.util.HashMap;
 
 import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.tm4e.core.internal.parser.PropertySettable;
-import org.eclipse.tm4e.core.theme.IRawTheme;
-import org.eclipse.tm4e.core.theme.IRawThemeSetting;
-import org.eclipse.tm4e.core.theme.IThemeSetting;
 
 public final class ThemeRaw extends HashMap<String, @Nullable Object>
 		implements IRawTheme, IRawThemeSetting, IThemeSetting, PropertySettable<Object> {
